README: update the README with project scope
This PR updates the README with the preoject scope as well as adding
a `super_linter` status badge. The status badge will signal if the
`linter` workflow is passing. It will show the last workflow run accross
all branches.

Additionally the `linter.yaml` was updated so the main branch can also
pass linting tests.